Move concrete implementation to Core
