+++ MvvmCross/Core/Core/Platform/LogProviders/LogMessageFormatter.cs
@@ -4,7 +4,7 @@ using System.Globalization;
 using System.Linq;
 using System.Text.RegularExpressions;
 
-namespace MvvmCross.Platform.Logging.LogProviders
+namespace MvvmCross.Core.Platform.LogProviders
 {
     internal static class LogMessageFormatter
     {
