+++ MvvmCross/Core/Core/Platform/LogProviders/NLogLogProvider.cs
@@ -3,8 +3,9 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
+using MvvmCross.Platform.Logging;
 
-namespace MvvmCross.Platform.Logging.LogProviders
+namespace MvvmCross.Core.Platform.LogProviders
 {
     internal class NLogLogProvider : MvxBaseLogProvider
     {
