+++ MvvmCross/Core/Core/Platform/LogProviders/SerilogLogProvider.cs
@@ -2,8 +2,9 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Linq.Expressions;
 using System.Reflection;
+using MvvmCross.Platform.Logging;
 
-namespace MvvmCross.Platform.Logging.LogProviders
+namespace MvvmCross.Core.Platform.LogProviders
 {
     internal class SerilogLogProvider : MvxBaseLogProvider
     {
