+++ MvvmCross/Platform/Platform/Logging/IMvxLog.cs
@@ -6,6 +6,4 @@ namespace MvvmCross.Platform.Logging
     {
         bool Log(MvxLogLevel logLevel, Func<string> messageFunc, Exception exception = null, params object[] formatParameters);
     }
-
-    delegate bool Logger(MvxLogLevel logLevel, Func<string> messageFunc, Exception exception = null, params object[] formatParameters);
 }
