--- MvvmCross/Platform/Platform/Platform/Logging/IMvxLog.cs
@@ -1,7 +1,0 @@
-﻿using System;
-namespace MvvmCross.Platform.Platform
-{
-    public interface IMvxLog
-    {
-    }
-}
